The refactored get_notifier and get_maintainer seem overkill and introduce indirection. Removed.
