+++ traits/observers/tests/test_has_traits_helpers.py
@@ -9,23 +9,11 @@
 # Thanks for using Enthought open source!
 
 import unittest
-from unittest import mock
 
 from traits.api import (
     Bool, HasTraits, List, Instance, Int, Property,
 )
 from traits.observers import _has_traits_helpers as helpers
-from traits.observers.events._trait_observer_event import TraitObserverEvent
-from traits.observers._testing import (
-    create_graph,
-    DummyObserver,
-    DummyNotifier,
-)
-from traits.trait_base import Uninitialized
-
-
-def default_dispatcher(handler, event):
-    handler(event)
 
 
 class Bar(HasTraits):
@@ -143,158 +131,3 @@ class TestHasTraitsHelpersIterObjects(unittest.TestCase):
         foo = Foo()
         list(helpers.iter_objects(foo, "property_value"))
         self.assertEqual(foo.property_n_calculations, 0)
-
-
-class TestHasTraitsHelpersGetNotifier(unittest.TestCase):
-    """ Test get_notifier. """
-
-    def setUp(self):
-        # holding strong references to handler and target.
-        self.handler = mock.Mock()
-        self.target = mock.Mock()
-
-        self.notifier = helpers.get_notifier(
-            handler=self.handler,
-            target=self.target,
-            dispatcher=default_dispatcher,
-        )
-
-    def test_get_notifier_good_change(self):
-        # when
-        bar = Bar()
-        self.notifier(
-            object=bar, name="value", old=0, new=1,
-        )
-
-        # then
-        self.assertEqual(self.handler.call_count, 1)
-        ((event, ), _), = self.handler.call_args_list
-        self.assertIsInstance(event, TraitObserverEvent)
-        self.assertIs(event.object, bar)
-        self.assertEqual(event.name, "value")
-        self.assertEqual(event.old, 0)
-        self.assertEqual(event.new, 1)
-
-    def test_get_notifier_old_uninitialized(self):
-        # when
-        self.notifier(
-            object=Bar(), name="value", old=Uninitialized, new=1,
-        )
-
-        # then
-        # the event is silenced.
-        self.assertEqual(self.handler.call_count, 0)
-
-
-class TestHasTraitsHelpersGetMaintainer(unittest.TestCase):
-    """ Test get_maintainer. """
-
-    def setUp(self):
-        # holding strong references to handler and target.
-        self.handler = mock.Mock()
-        self.target = mock.Mock()
-
-    def test_get_maintainer_changed_to_new_instance(self):
-        # Test the logic exercised by the maintainer.
-
-        class BarObserver(DummyObserver):
-            """ An observer specialized in observing Bar.value """
-
-            def iter_observables(self, object):
-                yield object._trait("value", 2)
-
-            def iter_objects(self, object):
-                raise NotImplementedError("Not used in this test.")
-
-        notifier = DummyNotifier()
-        bar_observer = BarObserver(notifier=notifier)
-
-        maintainer = helpers.get_maintainer(
-            graph=create_graph(bar_observer),
-            handler=self.handler,
-            target=self.target,
-            dispatcher=default_dispatcher,
-        )
-
-        # when
-        foo = Foo()
-        bar1 = Bar()
-        maintainer(
-            object=foo, name="instance", old=Uninitialized, new=bar1,
-        )
-
-        # then
-        self.assertEqual(
-            bar1._trait("value", 2)._notifiers(True),
-            [notifier],
-        )
-
-        # when
-        bar2 = Bar()
-        maintainer(
-            object=foo, name="instance", old=bar1, new=bar2,
-        )
-        # then
-        self.assertEqual(
-            bar1._trait("value", 2)._notifiers(True),
-            [],
-        )
-        self.assertEqual(
-            bar2._trait("value", 2)._notifiers(True),
-            [notifier],
-        )
-
-    def test_get_maintainer_changed_to_none(self):
-        # Instance may accept None, maintainer should accommodate that.
-
-        class UnassumingObserver(DummyObserver):
-            def iter_observables(self, object):
-                if object is None:
-                    raise ValueError("This observer cannot handle None.")
-                yield from ()
-
-            def iter_objects(self, object):
-                raise NotImplementedError("Not used in this test.")
-
-        maintainer = helpers.get_maintainer(
-            graph=create_graph(UnassumingObserver()),
-            handler=self.handler,
-            target=self.target,
-            dispatcher=default_dispatcher,
-        )
-
-        try:
-            maintainer(object=mock.Mock(), name="any", old=None, new=None)
-        except Exception:
-            self.fail("Maintainer should not fail when new value is None.")
-
-    def test_get_maintainer_excuse_old_value_with_no_notifiers(self):
-        # The "instance_with_default" trait has a default that has not been
-        # materialized prior to the user setting a new value to the trait.
-        # There isn't an old: Uninitialized -> new: Default value change event.
-        # Instead, there is a old: Default -> new value event.
-        # The old default value in this event won't have any notifiers
-        # to be removed, therefore we need to excuse the NotifierNotFound
-        # in the maintainer when it tries to remove notifiers from the old
-        # value.
-
-        class Observer(DummyObserver):
-            def iter_observables(self, object):
-                yield object
-
-        foo = ClassWithInstanceDefault()
-        maintainer = helpers.get_maintainer(
-            graph=create_graph(Observer()),
-            handler=self.handler,
-            target=self.target,
-            dispatcher=default_dispatcher,
-        )
-        ctrait = foo._trait("instance_with_default", 2)
-        ctrait._notifiers(True).append(maintainer)
-
-        try:
-            foo.instance_with_default = Bar()
-        except Exception:
-            self.fail(
-                "Reassigning the instance value should not fail."
-            )
